% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDP_weights.R
\name{get_GDP_weights}
\alias{get_GDP_weights}
\title{Get GDP Weights}
\usage{
get_GDP_weights(year)
}
\arguments{
\item{year}{Year for which weights are requested.}
}
\value{
A named vector of weights for the given year.
}
\description{
Retrieves the GDP weights for a specific year.
}
