% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_weights.R
\name{get_population_weights}
\alias{get_population_weights}
\title{Get Population Weights}
\usage{
get_population_weights(year)
}
\arguments{
\item{year}{Year for which weights are requested.}
}
\value{
A named vector of weights for the given year.
}
\description{
Retrieves the population weights for a specific year.
}
