% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responsibility_weights.R
\docType{data}
\name{regional_weights}
\alias{regional_weights}
\title{Regional Responsibility Weights}
\description{
This script defines responsibility weights for regions across years
and provides a function to retrieve weights for a specific year.
}
